requirements.txt
django-allauth==0.52.0
djangorestframework==3.12.4
goose3==3.1.12
gunicorn==20.1.0
langdetect==1.0.9
mysqlclient==2.1.0
nltk==3.6.5
python-dotenv==0.21.1
polyglot==16.7.4
pycryptodome==3.17
pycld2==0.41 
pyicu==2.10.2
requests==2.28.2
schedule==1.1.0
spacy==3.2.1
summarizer==0.0.7
